Refactor alertIcon in header into separate file
