+++ frontend/src/components/Header/Header.tsx
@@ -1,18 +1,15 @@
 import { config } from 'config'
-import { Button, Icon, TopBar, Typography, Popover } from '@equinor/eds-core-react'
+import { Button, Icon, TopBar, Typography } from '@equinor/eds-core-react'
 import { useInstallationContext } from 'components/Contexts/InstallationContext'
 import styled from 'styled-components'
-import { SelectLanguage } from './LanguageSelector'
+import { SelectLanguage } from 'components/Header/LanguageSelector'
 import { Icons } from 'utils/icons'
 import { useAlertContext } from 'components/Contexts/AlertContext'
-import { AlertListItem } from 'components/Alerts/AlertsListItem'
-import { useState, useRef } from 'react'
-import { tokens } from '@equinor/eds-tokens'
 import { AlertBanner } from 'components/Alerts/AlertsBanner'
-import { useLanguageContext } from 'components/Contexts/LanguageContext'
 import { FrontPageSectionId } from 'models/FrontPageSectionId'
 import { NavigationMenu } from 'components/NavigationMenu/NavigationMenu'
 import { findNavigationPage } from 'components/Pages/AssetSelectionPage/AssetSelectionPage'
+import { AlertIcon } from 'components/Header/AlertIcon'
 
 const StyledTopBar = styled(TopBar)`
     align-items: center;
@@ -52,25 +49,6 @@ const StyledTopBarHeader = styled.div`
     flex-direction: row;
     gap: 4px;
 `
-
-const StyledAlertPopoverHeader = styled(Popover.Header)`
-    width: 350px;
-`
-const StyledAlertPopoverTitle = styled(Popover.Title)`
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-right: 0.25em !important;
-`
-const Circle = styled.div`
-    position: absolute;
-    margin: 14px 23px 0px;
-    width: 9px;
-    height: 9px;
-    border-radius: 50%;
-`
-
 const StyledNavigationMenu = styled.div`
     display: none;
     @media (max-width: 600px) {
@@ -79,20 +57,8 @@ const StyledNavigationMenu = styled.div`
 `
 
 export const Header = ({ page }: { page: string }) => {
-    const { alerts, listAlerts } = useAlertContext()
+    const { alerts } = useAlertContext()
     const { installationName } = useInstallationContext()
-    const { TranslateText } = useLanguageContext()
-
-    const [isAlertDialogOpen, setIsAlertDialogOpen] = useState<boolean>(false)
-    const referenceElementNotifications = useRef<HTMLButtonElement>(null)
-
-    const onAlertOpen = () => {
-        setIsAlertDialogOpen(true)
-    }
-
-    const onAlertClose = () => {
-        setIsAlertDialogOpen(false)
-    }
 
     return (
         <>
@@ -117,18 +83,7 @@ export const Header = ({ page }: { page: string }) => {
                 </StyledWrapper>
                 <TopBar.Actions>
                     <IconStyle>
-                        <Button
-                            variant="ghost_icon"
-                            onClick={!isAlertDialogOpen ? onAlertOpen : onAlertClose}
-                            ref={referenceElementNotifications}
-                        >
-                            <Icon name={Icons.Notifications} size={24} />
-                            {Object.entries(listAlerts).length !== 0 &&
-                                installationName &&
-                                page !== 'root' && ( //Alert banners
-                                    <Circle style={{ background: tokens.colors.interactive.danger__resting.hex }} />
-                                )}
-                        </Button>
+                        {page !== 'root' && <AlertIcon />}
                         <Button
                             variant="ghost_icon"
                             onClick={() => {
@@ -141,39 +96,6 @@ export const Header = ({ page }: { page: string }) => {
                     <SelectLanguageWrapper>{SelectLanguage()}</SelectLanguageWrapper>
                 </TopBar.Actions>
             </StyledTopBar>
-            <Popover
-                onClose={onAlertClose}
-                open={isAlertDialogOpen}
-                placement={'bottom'}
-                anchorEl={referenceElementNotifications.current}
-            >
-                <StyledAlertPopoverHeader>
-                    <StyledAlertPopoverTitle>
-                        <Typography variant="h6">{TranslateText('Alerts')}</Typography>
-                        <Button variant={'ghost_icon'} onClick={onAlertClose}>
-                            <Icon name="close" size={24} />
-                        </Button>
-                    </StyledAlertPopoverTitle>
-                </StyledAlertPopoverHeader>
-                <Popover.Content>
-                    {Object.entries(listAlerts).length === 0 && installationName && page !== 'root' && (
-                        <Typography variant="h6">{TranslateText('No alerts')}</Typography>
-                    )}
-                    {Object.entries(listAlerts).length > 0 && installationName && page !== 'root' && (
-                        <StyledAlertList>
-                            {Object.entries(listAlerts).map(([key, value]) => (
-                                <AlertListItem
-                                    key={key}
-                                    dismissAlert={value.dismissFunction}
-                                    alertCategory={value.alertCategory}
-                                >
-                                    {value.content}
-                                </AlertListItem>
-                            ))}
-                        </StyledAlertList>
-                    )}
-                </Popover.Content>
-            </Popover>
             <StyledNavigationMenu>
                 <NavigationMenu />
             </StyledNavigationMenu>
